requirements.txt
openai
langchain
python-dotenv
azure-ai-openai